bit of refactoring
